producer/requirements.txt
python-dotenv
requests
confluent-kafka
